Reorganize structure to be a lot easier to work with.
