+++ src/root/bonfire/actions/actions.component.scss
@@ -1,12 +1,7 @@
-.base-container {
-  margin: 0 24px;
-}
-
 .tab-container {
   margin: 12px 50%;
   color: white;
 }
-
 
 ::ng-deep .mat-tab-list .mat-tab-label {
   color:lightgray;
