+++ src/root/bonfire/bonfire.component.scss
@@ -1,0 +1,23 @@
+.base-container {
+  margin: 0 24px;
+}
+
+.tab-container {
+  margin: 12px 50%;
+  color: white;
+}
+
+
+::ng-deep .mat-tab-list .mat-tab-label {
+  color:lightgray;
+}
+
+.actions {
+  display: flex;
+  flex-flow: row;
+  margin: 5px 5px 5px 5px;
+  justify-content: center;
+  .action-button {
+    margin: 5px 10px;
+  }
+}
