+++ tslint.json
@@ -16,7 +16,7 @@
       "camelCase"
     ],
     "component-selector": [
-      true,
+      false,
       "element",
       "",
       "kebab-case"
